Before:

```bash
$ go test -bench=. -benchmem -run=^$
goarch: amd64
pkg: github.com/masslessparticle/go-framestruct
cpu: Intel(R) Core(TM) i5-8500 CPU @ 3.00GHz
BenchmarkMarshalTable_10/marshal-6                126786              9355 ns/op            3104 B/op         99 allocs/op
BenchmarkMarshalTable_100/marshal-6                16560             72396 ns/op           20336 B/op        648 allocs/op
BenchmarkMarshalTable_1000/marshal-6                1735            690775 ns/op          171152 B/op       6057 allocs/op
BenchmarkMarshalTable_10000/marshal-6                165           7195916 ns/op         3198808 B/op      60084 allocs/op
BenchmarkEmbeddedStruct_10/marshal-6              103455             11504 ns/op            3000 B/op        147 allocs/op
BenchmarkEmbeddedStruct_100/marshal-6              10000            103790 ns/op           22781 B/op       1233 allocs/op
BenchmarkEmbeddedStruct_1000/marshal-6              1159           1017648 ns/op          209837 B/op      12039 allocs/op
BenchmarkEmbeddedStruct_10000/marshal-6              100          10448728 ns/op         2813916 B/op     120061 allocs/op
PASS
ok      github.com/masslessparticle/go-framestruct      11.337s
```

After:

```bash
go test -bench=. -benchmem -run=^$
goos: darwin
goarch: amd64
pkg: github.com/masslessparticle/go-framestruct
cpu: Intel(R) Core(TM) i5-8500 CPU @ 3.00GHz
BenchmarkMarshalTable_10/marshal-6                152192              7771 ns/op            3104 B/op         99 allocs/op
BenchmarkMarshalTable_100/marshal-6                20836             58110 ns/op           20336 B/op        648 allocs/op
BenchmarkMarshalTable_1000/marshal-6                2067            549002 ns/op          171152 B/op       6057 allocs/op
BenchmarkMarshalTable_10000/marshal-6                199           5942379 ns/op         3198814 B/op      60084 allocs/op
BenchmarkEmbeddedStruct_10/marshal-6              158178              7455 ns/op            2360 B/op        107 allocs/op
BenchmarkEmbeddedStruct_100/marshal-6              19177             62640 ns/op           16376 B/op        833 allocs/op
BenchmarkEmbeddedStruct_1000/marshal-6              1994            603470 ns/op          145785 B/op       8039 allocs/op
BenchmarkEmbeddedStruct_10000/marshal-6              193           6203007 ns/op         2172929 B/op      80057 allocs/op
PASS
ok      github.com/masslessparticle/go-framestruct      12.337s

```